i40iw: remove support for ib_get_vector_affinity
devices that does not use managed affinity can not export a vector
affinity as the consumer relies on having a static mapping it can map
to upper layer affinity (e.g. sw queues). If the driver allows the user
to set the device irq affinity, then the affinitization of a long term
existing entites is not relevant.

For example, nvme-rdma controllers queue-irq affinitization is determined
at init time so if the irq affinity changes over time, we are no longer
aligned.

Signed-off-by: Sagi Grimberg <sagi@grimberg.me>
Acked-by: Shiraz Saleem <shiraz.saleem@intel.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>